リファクタリング（games_controller）
